make all the CRUD private

+++ mycontact-backend/controllers/contactController.js
@@ -3,15 +3,15 @@ const Contact = require('../models/contactModule')
 
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find()
+    const contacts = await Contact.find({ user_id: req.user.id })
     res.status(200).json(contacts)
 })
 
 //@desc Get a contact
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
 
     const contact = await Contact.findById(req.params.id)
@@ -26,7 +26,7 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@desc Create a contact
 //@route POST /api/contacts
-//@access public
+//@access private
 //201 = resource created
 const createContact = asyncHandler(async (req, res) => {
     console.log("The request body is : ", req.body);
@@ -40,7 +40,8 @@ const createContact = asyncHandler(async (req, res) => {
     const contact = await Contact.create({
         name,
         email,
-        phone
+        phone,
+        user_id: req.user.id, 
     })
 
     res.status(201).json(contact)
@@ -48,13 +49,18 @@ const createContact = asyncHandler(async (req, res) => {
 
 //@desc Update a contact
 //@route PUT /api/contacts
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id)
 
     if (!contact) {
         res.status(404)
         throw new Error("Contact not found")
+    }
+
+    if(contact.user_id.toString() !== req.user.id ){
+        res.status(401)
+        throw new Error ("User don't have permission to update other user contacts")
     }
 
     const updatedContact = await Contact.findByIdAndUpdate(
@@ -68,11 +74,23 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@desc delete a contact
 //@route DELETE /api/contacts
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
 
-    const contact = await Contact.findOneAndDelete({_id: req.params.id});
-    res.status(200).json(contact);
+    const contact = await Contact.findById(req.params.id)
+
+    if (!contact) {
+        res.status(404)
+        throw new Error("Contact not found")
+    }
+
+    if(contact.user_id.toString() !== req.user.id ){
+        res.status(401)
+        throw new Error ("User don't have permission to update other user contacts")
+    }
+
+    const updatedContact = await Contact.findOneAndDelete({_id: req.params.id});
+    res.status(200).json(updatedContact);
 
 })
